refactor some of the classes to add potential
