+++ cogsworth/fire/__init__.py
@@ -1,1 +1,1 @@
-from . import fire, readsnap, centre, utils
+from . import centre, pop, snap, potential, utils
